change structure of project accordance with DI of automapping
